service/main_test.go
// File: cmd/main_test.go
package main

import (
	"context"
	"errors"
	"notes-server/internal/server"
	"testing"
	"time"

	"github.com/kardianos/service"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
)

// MockService implements service.Service interface for testing
type MockService struct {
	mock.Mock
}

func (m *MockService) Run() error {
	args := m.Called()
	return args.Error(0)
}

func (m *MockService) Start() error {
	args := m.Called()
	return args.Error(0)
}

func (m *MockService) Stop() error {
	args := m.Called()
	return args.Error(0)
}

func (m *MockService) Restart() error {
	args := m.Called()
	return args.Error(0)
}

func (m *MockService) Install() error {
	args := m.Called()
	return args.Error(0)
}

func (m *MockService) Uninstall() error {
	args := m.Called()
	return args.Error(0)
}

func (m *MockService) Status() (service.Status, error) {
	args := m.Called()
	return args.Get(0).(service.Status), args.Error(1)
}

func (m *MockService) Logger(errs chan<- error) (service.Logger, error) {
	args := m.Called(errs)
	return args.Get(0).(service.Logger), args.Error(1)
}

func (m *MockService) SystemLogger(errs chan<- error) (service.Logger, error) {
	args := m.Called(errs)
	return args.Get(0).(service.Logger), args.Error(1)
}

func (m *MockService) String() string {
	args := m.Called()
	return args.String(0)
}

func (m *MockService) Platform() string {
	args := m.Called()
	return args.String(0)
}

// MockLogger implements service.Logger interface for testing
type MockLogger struct {
	mock.Mock
}

func (m *MockLogger) Error(v ...interface{}) error {
	args := m.Called(v...)
	return args.Error(0)
}

func (m *MockLogger) Warning(v ...interface{}) error {
	args := m.Called(v...)
	return args.Error(0)
}

func (m *MockLogger) Info(v ...interface{}) error {
	args := m.Called(v...)
	return args.Error(0)
}

func (m *MockLogger) Errorf(format string, a ...interface{}) error {
	args := m.Called(format, a)
	return args.Error(0)
}

func (m *MockLogger) Warningf(format string, a ...interface{}) error {
	args := m.Called(format, a)
	return args.Error(0)
}

func (m *MockLogger) Infof(format string, a ...interface{}) error {
	args := m.Called(format, a)
	return args.Error(0)
}

// TestHandleServiceCommand tests all service commands
func TestHandleServiceCommand(t *testing.T) {
	tests := []struct {
		name        string
		command     string
		setupMock   func(*MockService)
		expectError bool
	}{
		{
			name:    "install successful",
			command: "install",
			setupMock: func(m *MockService) {
				m.On("Install").Return(nil)
			},
			expectError: false,
		},
		{
			name:    "install fails",
			command: "install",
			setupMock: func(m *MockService) {
				m.On("Install").Return(errors.New("install failed"))
			},
			expectError: true,
		},
		{
			name:    "start successful",
			command: "start",
			setupMock: func(m *MockService) {
				m.On("Start").Return(nil)
			},
			expectError: false,
		},
		{
			name:    "stop successful",
			command: "stop",
			setupMock: func(m *MockService) {
				m.On("Stop").Return(nil)
			},
			expectError: false,
		},
		{
			name:    "restart successful",
			command: "restart",
			setupMock: func(m *MockService) {
				m.On("Restart").Return(nil)
			},
			expectError: false,
		},
		{
			name:    "status running",
			command: "status",
			setupMock: func(m *MockService) {
				m.On("Status").Return(service.StatusRunning, nil)
			},
			expectError: false,
		},
		{
			name:    "uninstall successful",
			command: "uninstall",
			setupMock: func(m *MockService) {
				m.On("Uninstall").Return(nil)
			},
			expectError: false,
		},
		{
			name:    "invalid command",
			command: "invalid",
			setupMock: func(m *MockService) {},
			expectError: true,
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			mockSvc := &MockService{}
			tt.setupMock(mockSvc)

			err := handleServiceCommand(mockSvc, tt.command)
			if tt.expectError {
				assert.Error(t, err)
			} else {
				assert.NoError(t, err)
			}
		})
	}
}

// TestProgram tests the program struct implementation
func TestProgram(t *testing.T) {
	ctx, cancel := context.WithCancel(context.Background())
	defer cancel()

	mockLogger := &MockLogger{}
	mockLogger.On("Info", mock.Anything).Return(nil)
	logger = mockLogger

	srv := server.NewServer("test-server")
	p := &program{
		srv:    srv,
		ctx:    ctx,
		cancel: cancel,
	}

	// Test Start
	mockSvc := &MockService{}
	err := p.Start(mockSvc)
	assert.NoError(t, err)

	// Give some time for the goroutine to start
	time.Sleep(100 * time.Millisecond)

	// Test Stop
	err = p.Stop(mockSvc)
	assert.NoError(t, err)

	// Verify context was cancelled
	select {
	case <-p.ctx.Done():
		// Context was cancelled as expected
	default:
		t.Error("Context was not cancelled")
	}
}

// TestMain_NoArgs tests the main function without arguments
func TestMain_NoArgs(t *testing.T) {
	t.Skip("Skipping main test as it requires special environment setup")
}